feat:  massive improvements in endpoints

+++ src/app/api/themes/delete-theme/route.ts
@@ -2,33 +2,48 @@ import { RATE_LIMIT_CONFIGS, rateLimitter } from "@/middleware/rate-limit";
 import { NextResponse, NextRequest } from "next/server";
 import { getUserToken } from "@/utils/services/auth/getToken";
 
+export const runtime = 'edge';
 
-export const deleteTheme = async (req: NextRequest) => {
-
+const deleteTheme = async (req: NextRequest) => {
     const API_ENDPOINT = process.env.SERVER_DELETE_THEME;
-    const token = await getUserToken();
 
-    const themeID = await req.json()
-    console.log("Received theme ID to delete:", themeID.themeID);
-
-    const response = await fetch(`${API_ENDPOINT}${themeID.themeID}`,
-        {
-            method: "DELETE",
-            headers: {
-                "Content-Type": "application/json",
-                "Authorization": `Bearer ${token}`,
-            },
-        }
-    );
-
-    if (!response.ok) {
-        throw new Error("Error al realizar la petición");
+    if (!API_ENDPOINT) {
+        return NextResponse.json(
+            { error: "API endpoint no está definido" },
+            { status: 500 }
+        );
     }
 
-    const data = await response.json();
-    console.log("Theme deleted successfully:", data.message);
-    return NextResponse.json({ message: data.message });
+    try {
+        const token = await getUserToken();
+        const themeID = await req.json()
+        const response = await fetch(`${API_ENDPOINT}${themeID.themeID}`,
+            {
+                method: "DELETE",
+                headers: {
+                    "Content-Type": "application/json",
+                    "Authorization": `Bearer ${token}`,
+                },
+                signal: AbortSignal.timeout(5000)
+            }
+        );
 
+        if (!response.ok) {
+            const errorData = await response.json();
+            throw new Error(errorData.details || "Error al eliminar el tema");
+        }
+
+        const data = await response.json();
+        return NextResponse.json({ message: data.message });
+
+
+    } catch (error) {
+        console.error({ error: error });
+        return NextResponse.json({ error: "error al eliminar el tema" }, { status: 500 });
+    }
 }
 
-export const DELETE = rateLimitter({ fn: deleteTheme, options: RATE_LIMIT_CONFIGS.WRITE });
+export const DELETE = rateLimitter({
+    fn: deleteTheme,
+    options: { ...RATE_LIMIT_CONFIGS.WRITE, identifier: 'deleteTheme' }
+});
